Rename to reflectometry server
